+++ engine/entity.go
@@ -7,7 +7,7 @@ import (
 // Entity is a thing in the game that can be updated and drawn.
 type Entity interface {
 	// Absolute time for animations; dt for physics.
-	Update(frameState *UpdateState, dt float64)
+	Update(frameState *UpdateState, dt float32)
 	Draw(*Graphics, *Camera)
 
 	Impulse(*Vec2)
